refactor: Mapper 파일 제거 및 데이터 클래스에 통합으로 인한 코드 수정 #68
