+++ local/src/main/java/com/kdjj/local/dataSource/RecipeLocalDataSourceImpl.kt
@@ -3,7 +3,7 @@ package com.kdjj.local.dataSource
 import com.kdjj.data.recipe.RecipeLocalDataSource
 import com.kdjj.domain.model.Recipe
 import com.kdjj.local.dao.RecipeDao
-import com.kdjj.local.toEntity
+import com.kdjj.local.model.toEntity
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
 import javax.inject.Inject
