+++ local/src/main/java/com/kdjj/local/dataSource/RecipeTypeLocalDataSourceImpl.kt
@@ -3,8 +3,8 @@ package com.kdjj.local.dataSource
 import com.kdjj.data.recipetype.RecipeTypeLocalDataSource
 import com.kdjj.domain.model.RecipeType
 import com.kdjj.local.dao.RecipeTypeDao
-import com.kdjj.local.toDomain
-import com.kdjj.local.toEntity
+import com.kdjj.local.model.toDomain
+import com.kdjj.local.model.toEntity
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
 import javax.inject.Inject
