+++ remote/src/main/java/com/kdjj/remote/FirestoreDaoImpl.kt
@@ -4,6 +4,7 @@ import com.google.firebase.firestore.FirebaseFirestore
 import com.google.firebase.firestore.ktx.toObject
 import com.kdjj.domain.model.RecipeType
 import com.kdjj.remote.entity.RecipeTypeEntity
+import com.kdjj.remote.entity.toDomain
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.tasks.await
 import kotlinx.coroutines.withContext
