+++ remote/src/main/java/com/kdjj/remote/entity/RecipeStepEntity.kt
@@ -1,12 +1,23 @@
 package com.kdjj.remote.entity
 
+import com.kdjj.domain.model.RecipeStep
 import com.kdjj.domain.model.RecipeStepType
 
-data class RecipeStepEntity (
-    val stepId: String = "",
-    val name: String = "",
-    val type: RecipeStepType = RecipeStepType.FRY,
-    val description: String = "",
-    val imgPath: String = "",
-    val seconds: Int = 0
+internal data class RecipeStepEntity(
+	val stepId: String = "",
+	val name: String = "",
+	val type: RecipeStepType = RecipeStepType.FRY,
+	val description: String = "",
+	val imgPath: String = "",
+	val seconds: Int = 0
 )
+
+internal fun RecipeStepEntity.toDomain(): RecipeStep =
+	RecipeStep(
+		stepId,
+		name,
+		type,
+		description,
+		imgPath,
+		seconds
+	)
