+++ remote/src/main/java/com/kdjj/remote/entity/RecipeTypeEntity.kt
@@ -1,6 +1,11 @@
 package com.kdjj.remote.entity
 
-data class RecipeTypeEntity(
+import com.kdjj.domain.model.RecipeType
+
+internal data class RecipeTypeEntity(
 	val id: Int = 0,
 	val title: String = ""
 )
+
+internal fun RecipeTypeEntity.toDomain(): RecipeType =
+	RecipeType(id, title)
